requirements.txt
scapy
python-nmap
mac-vendor-lookup
psutil
pandas

